Move all softints to threads.
This is a temporary implementation and it sucks ballz since we're
using the SCHED_LOCK as an interlock to "keep state on a wakeup", in
the future this should all be rewritten anyway. I need this to hold
everything together in order to remove the IPL subsystem.

There are still 2 or 3 devices registering their own softints which I
did not touch.