Unify ithread_softmain() into ithread().
Had to make softhandlers return 1 so we don't the the stray interrupt
report.

Also had to give an empty hwmask/unmask implementation to softpic.

While here, make is_scheduled volatile since it might be changed by an
interrupt handler right in between invocations. This _only_ prevents
the compiler from keeping each read on register, which is what we
want.