Hardclock work for critical sections.
We need to block hardclock when on a critical section and resume it on
critical exit, for that we need a fake trapframe, openbsd was using a
very long frame, clock actually only needs 5 registers, so split them up
and use a smaller frame for Xfakeclock().

ci->ci_req will be a bitmap, and clock should be only one bit.