Make so that mi_switch() won't relock sched_lock.
The caller needs to reacquire it if desired. More complex paths like
proc_stop() were left like before, so they reacquire the lock.

Almost every user of mi_switch() just do a SCHED_UNLOCK() after
return, so this commit prevents a LOCK/UNLOCK on a contented lock.

For this to work, you can never mi_switch with a recursive sched_lock,
and this is asserted.